Refactoring, support for newer images

+++ install/setup.sh
@@ -1,10 +1,12 @@
+echo "Setup at launch time (nricheton/jeedom-optimized)"
+
 if [ ! -z ${APACHE_PORT} ]; then
 	echo 'Change apache listen port to: '${APACHE_PORT}
 	echo "Listen ${APACHE_PORT}" > /etc/apache2/ports.conf
 	sed -i -E "s/\<VirtualHost \*:(.*)\>/VirtualHost \*:${APACHE_PORT}/" /etc/apache2/sites-enabled/000-default.conf
-else
-	echo "Listen 80" > /etc/apache2/ports.conf
-	sed -i -E "s/\<VirtualHost \*:(.*)\>/VirtualHost \*:80/" /etc/apache2/sites-enabled/000-default.conf
+#else
+#	echo "Listen 80" > /etc/apache2/ports.conf
+#	sed -i -E "s/\<VirtualHost \*:(.*)\>/VirtualHost \*:80/" /etc/apache2/sites-enabled/000-default.conf
 fi
 
 
@@ -23,11 +25,18 @@ if [ ! -z ${HOSTNAME} ]; then
     echo ":1 ${HOSTNAME}" >> /etc/hosts
 fi
 
+REMOVE_MARIADB=false
 # Remove mariadb in case some plugin adds it 
-apt-get remove -y mariadb-client mariadb-common mariadb-server
+if [ ${REMOVE_MARIADB} = true ]; then
+  echo "Remove mariadb"
+  apt-get remove -y mariadb-client mariadb-common mariadb-server
+fi
 
-# Ensure RF link works with a recent node version
-cd /var/www/html/plugins/rflink/resources && npm rebuild && npm install && chown -R www-data node_modules
+INSTALL_RFLINK=false
+if [ ${INSTALL_RFLINK} = true ]; then
+	# Ensure RF link works with a recent node version
+	cd /var/www/html/plugins/rflink/resources && npm rebuild && npm install && chown -R www-data node_modules
+fi
 
 # Ignore error from previous command
-echo "Setup ok"
+echo "Setup at launch time : done"
